correction conflits

+++ src/main/java/SopraAJC/NotreProjet/restcontroller/SessionBatimentRestController.java
@@ -106,13 +106,13 @@ public class SessionBatimentRestController {
 	
 	@GetMapping("/transformation/{idPartie}/{idCompte}")
 	@JsonView(JsonViews.SessionBatimentWithBatiment.class)
-	public List<SessionBatiment> batimentTransformation(@PathVariable Integer idPartie, @PathVariable Integer idCompte){
+	public List<SessionBatiment> batimentTransformation(@PathVariable Long idPartie, @PathVariable Long idCompte){
 		return gestionRessource.listeBatimentTransformation(sessionRepo.findByPartieAndCompte(partieRepo.findById(idPartie).get(), compteRepo.findById(idCompte).get()).get());
 	}
 
 	@PostMapping("/construction/{idPartie}/{idCompte}/{idBat}")
 	@JsonView(JsonViews.SessionBatimentWithBatiment.class)
-	public SessionBatiment construction(@PathVariable Integer idPartie,@PathVariable Integer idCompte, @PathVariable Integer idBat) {
+	public SessionBatiment construction(@PathVariable Long idPartie,@PathVariable Long idCompte, @PathVariable Integer idBat) {
 		
 		if(construction.verificationConstructible(batimentRepo.findById(idBat).get(), sessionRepo.findByPartieAndCompte(partieRepo.findById(idPartie).get(), compteRepo.findById(idCompte).get()).get())) {
 			SessionBatiment sb= construction.constructBat(batimentRepo.findById(idBat).get(), sessionRepo.findByPartieAndCompte(partieRepo.findById(idPartie).get(), compteRepo.findById(idCompte).get()).get());
